+++ src/main.ts
@@ -1,4 +1,4 @@
 import './style.css';
-import { RegistrationPage } from './pages/registration/registration';
+import { RegistrationPage } from './pages/registration/layout';
 
 RegistrationPage();
