+++ src/pages/registration/layout.ts
